back/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==30.8.2
gunicorn==21.2.0
mpmath==1.3.0
pillow==11.0.0
PyJWT==2.10.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.3
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.2
dj-database-url==2.1.0
python-dotenv==1.0.0
whitenoise==6.6.0